% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format.R
\name{check_format}
\alias{check_format}
\title{check_format()}
\usage{
check_format(mabac_df)
}
\arguments{
\item{mabac_df}{The original dataset to be validated}
}
\value{
TRUE if everything is correct or an error containing the explanatory message.
}
\description{
Verify if the database is correctly formatted
}
