% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabacR.R
\name{mabacR}
\alias{mabacR}
\title{mabacR()}
\usage{
mabacR(mabac_df)
}
\arguments{
\item{mabac_df}{The original dataset containing, necessarily in this order:
1. a column with the criteria;
2. a column with the weights (which must add up to 1);
3. A column with types (1 for maximization and -1 for minimization);
4. and the columns with the evaluated items.}
}
\value{
A ranking with the optimal choice in accordance with your criteria, weights and type of criteria
}
\description{
This package will execute the mathematical algorithms that make up the MABAC method,
             a multi-criteria decision method. The data will be normalized,
             weighted by the weights, the border area will be determined,
             the distances to this border area will be calculated
             and finally a ranking with the optimal option will be generated.
}
\examples{
data(mabac_df)
          output <- mabacR(mabac_df)

}
