% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format.R
\name{check_format}
\alias{check_format}
\title{check_format()}
\usage{
check_format(mabac_df, verbose = TRUE)
}
\arguments{
\item{mabac_df}{The original dataset to be validated}

\item{verbose}{Logical. If \code{TRUE}, messages about errors and warnings will be printed to the console.}
}
\value{
A list containing:
\item{status}{Logical, \code{TRUE} if the format is correct, \code{FALSE} if not.}
\item{message}{A message providing details about any format issues.}
}
\description{
Verify if the database is correctly formatted
}
